接口测试完成

+++ chatbot-api-domain/src/main/java/com/chatbot/ai/domain/bilibili/model/vo/At_details.java
@@ -5,9 +5,9 @@ import lombok.Data;
 @Data
 public class At_details {
 
-    private Integer mid;
+    private Long mid;
 
-    private Integer fans;
+    private Long fans;
 
     private String nickname;
